+++ stop_booklab_server.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-#
-# stop BookLab HTTP server
-#
-
-# need to save PID of python httm small server on start it 
-# in order to know wjat process to kill
-