+++ CHANGELOG.md
@@ -37,7 +37,7 @@ First booklabd as Flask web app to serve `/api/.../` routes
 
 
 ## |... 0.4b2] - ...
-Create `db...` objects and a test `/api/bcat` route:
+Create database (aka `db*`) objects for JSON files and a test `/api/bcat` route:
 
 * ...tbd a test route for /api/bcat/ to write sone data in books catalog
 * ...tbd init the `db` objects to JSON databases (.../booklabd/data/)